extract PreferencesBasedStorageHandler from StorageHandlerImpl
the abstract class PreferencesBasedStorageHandler can easely be subclassed by a client to use another means than Gson for serialization
